Remove the logic for migrating from v0 to v1 and replace it with migrating from v0 to v2.
Since we will use `replace_all`, we should no longer use the `by_subject` key from v1.
