+++ apps/remote_control/lib/lexical/remote_control/search/store/backends/ets/schemas/v1.ex
@@ -8,8 +8,6 @@ defmodule Lexical.RemoteControl.Search.Store.Backends.Ets.Schemas.V1 do
   Finally, entries are stored by their reference, which powers direct lookups, which are used in fuzzy matching.
 
   """
-
-  alias Lexical.RemoteControl.Search.Indexer.Entry
   alias Lexical.RemoteControl.Search.Store.Backends.Ets.Schema
 
   use Schema, version: 1
@@ -28,68 +26,6 @@ defmodule Lexical.RemoteControl.Search.Store.Backends.Ets.Schemas.V1 do
   defkey :by_path, [:path]
 
   def migrate(entries) do
-    migrated =
-      entries
-      |> Stream.filter(fn
-        {_, %_{elixir_version: _, erlang_version: _, type: _, subtype: _, ref: _}} -> true
-        _ -> false
-      end)
-      |> Stream.map(fn {_, entry} -> entry end)
-      |> entries_to_rows()
-
-    {:ok, migrated}
+    {:ok, entries}
   end
-
-  @spec entries_to_rows(Enumerable.t(Entry.t())) :: [tuple()]
-  def entries_to_rows(entries) do
-    entries
-    |> Stream.flat_map(&to_rows(&1))
-    |> Enum.reduce(%{}, fn {key, value}, acc ->
-      Map.update(acc, key, [value], fn old_values -> [value | old_values] end)
-    end)
-    |> Enum.to_list()
-  end
-
-  def to_rows(%Entry{} = entry) do
-    subject_key =
-      by_subject(
-        elixir_version: entry.elixir_version,
-        erlang_version: entry.erlang_version,
-        subject: to_subject(entry.subject),
-        type: entry.type,
-        subtype: entry.subtype,
-        path: entry.path
-      )
-
-    id_key =
-      by_id(
-        id: entry.ref,
-        type: entry.type,
-        subtype: entry.subtype,
-        elixir_version: entry.elixir_version,
-        erlang_version: entry.erlang_version
-      )
-
-    path_key = by_path(path: entry.path)
-
-    [{subject_key, entry}, {id_key, entry}, {path_key, id_key}]
-  end
-
-  # This case will handle any namespaced entries
-  def to_rows(%{elixir_version: _, erlang_version: _, type: _, subtype: _, ref: _} = entry) do
-    map = Map.delete(entry, :__struct__)
-
-    Entry
-    |> struct(map)
-    |> to_rows()
-  end
-
-  def table_options do
-    [:named_table, :ordered_set]
-  end
-
-  defp to_subject(binary) when is_binary(binary), do: binary
-  defp to_subject(:_), do: :_
-  defp to_subject(atom) when is_atom(atom), do: inspect(atom)
-  defp to_subject(other), do: to_string(other)
 end
